requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113
torchvision==0.12.0+cu113
matplotlib
numpy
scipy
tqdm
avalanche-lib
avalanche-lib[extra]
avalanche-lib[detection]
